+++ GenericTelemetryProvider/RBRTelemetryProvider.cs
@@ -176,10 +176,10 @@ namespace GenericTelemetryProvider
 
         public override void SimulateEngine()
         {
-            rawData.max_rpm = 6000;
-            rawData.max_gears = 6;
-            rawData.gear = telemetryData.control_.gear_;
-            rawData.idle_rpm = 700;
+            rawData.max_rpm = 6000.0f;
+            rawData.max_gears = 6.0f;
+            rawData.gear = (float)telemetryData.control_.gear_;
+            rawData.idle_rpm = 700.0f;
             filteredData.engine_rate = telemetryData.car_.engine_.rpm_;
 
             Vector3 localVelocity = new Vector3((float)filteredData.local_velocity_x, (float)filteredData.local_velocity_y, (float)filteredData.local_velocity_z);
